requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/test.in
#
coverage==4.4.1           # via pytest-cov
funcsigs==1.0.2           # via mock
mock==2.0.0
pbr==3.1.1                # via mock
py==1.4.34                # via pytest, pytest-catchlog
pytest-catchlog==1.2.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.2.3             # via pytest-catchlog, pytest-cov, pytest-django
six==1.11.0               # via mock
